Seperate teaching_mode.py and motion_manager.py
